requirements.txt
python-dotenv>=0.19.0
pyspark>=3.3.0
requests>=2.28.1
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0